requirements.txt
uagents
uagents-ai-engine
langchain-google-genai
pillow
jupyter
python-dotenv
